requirements.txt
matplotlib==3.7.1
numpy==1.23.5
pandas==2.3.0
PyYAML==6.0.2
seaborn==0.13.2
torch==2.6.0
yfinance==0.2.63
